Further reverts

+++ src/main/java/com/mojang/minecraft/sound/Audio.java
@@ -1,5 +1,5 @@
 package com.mojang.minecraft.sound;
 
 public interface Audio {
-    boolean play(short[] var1, short[] var2, int var3);
+    boolean play(int[] var1, int[] var2, int var3);
 }
